absenvttv3-main/requirements.txt
asgiref==3.4.1
attrs==21.2.0
authentication==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.18.0
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.7.0
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.1
pycryptodome==3.10.1
PyJWT==2.1.0
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
whitenoise==5.3.0
WTForms==2.3.3
zipp==3.5.0
